move test log handler to test common
