+++ sdk/test/metrics/BUILD
@@ -47,6 +47,7 @@ cc_test(
     ],
     deps = [
         "metrics_common_test_utils",
+        "//test_common:headers",
         "@com_google_googletest//:gtest_main",
     ],
 )
